requirements.txt

numpy
matplotlib
pillow
pandas
